+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @elog/cli
 
+## 0.4.3
+
+### Patch Changes
+
+- 修复 Front Matter 中字符超长问题
+- dc21449: 修复 Front Matter 中字符超长问题
+- Updated dependencies
+- Updated dependencies [dc21449]
+  - @elog/core@0.4.3
+  - @elog/shared@0.4.3
+
+## 0.4.3-beta.0
+
+### Patch Changes
+
+- 修复 Front Matter 中字符超长问题
+- Updated dependencies
+  - @elog/core@0.4.3-beta.0
+  - @elog/shared@0.4.3-beta.0
+
 ## 0.4.2
 
 ### Patch Changes